# src/docs/roles/swagger-role-gm.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "GIRIK API - GM",
    "version": "1.0.0",
    "description": "APIs for **GM** role. Sections are grouped by module (Users, Clients, Vessels, Jobs, etc.). Use `POST /auth/login` first and set the token in Authorize."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local"
    },
    {
      "url": "https://api.girik.com/api/v1",
      "description": "Production"
    }
  ],
...
  "paths": {
...
    "/jobs": {
      "post": {
        "tags": [
          "Jobs"
        ],
        "summary": "Create Job",
        "responses": {
          "200": {
            "description": "Success"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "vessel_id",
                  "certificate_type_id"
                ],
                "properties": {
                  "vessel_id": {
                    "type": "string"
                  },
                  "certificate_type_id": {
                    "type": "string"
                  },
                  "reason": {
                    "type": "string"
                  },
                  "target_port": {
                    "type": "string"
                  },
                  "target_date": {
                    "type": "string",
                    "format": "date"
                  }
                }
              },
              "example": {
                "vessel_id": "{{vessel_id}}",
                "certificate_type_id": "{{certificate_type_id}}",
                "reason": "Annual survey and class renewal",
                "target_port": "Port of Singapore",
                "target_date": "2025-12-15"
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "List Jobs",
        "responses": {
          "200": {
            "description": "Success"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/jobs/{id}/assign": {
      "put": {
        "tags": [
          "Jobs"
        ],
        "summary": "Assign Surveyor",
        "responses": {
          "200": {
            "description": "Success"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "id"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "surveyorId"
                ],
                "properties": {
                  "surveyorId": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "surveyorId": "{{surveyor_id}}"
              }
            }
          }
        }
      }
    },
    "/jobs/{id}/status": {
      "put": {
        "tags": [
          "Jobs"
        ],
        "summary": "Update Job Status",
        "responses": {
          "200": {
            "description": "Success"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "id"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "PENDING",
                      "ASSIGNED",
                      "IN_PROGRESS",
                      "COMPLETED",
                      "CANCELLED"
                    ]
                  },
                  "remarks": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "status": "ASSIGNED",
                "remarks": "Job assigned to surveyor"
              }
            }
          }
        }
      }
    },
    "/jobs/{id}/cancel": {
      "put": {
        "tags": [
          "Jobs"
        ],
        "summary": "Cancel Job",
        "responses": {
          "200": {
            "description": "Success"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "id"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "reason": "Client requested cancellation"
              }
            }
          }
        }
      }
    },
    "/jobs/{id}/hold": {
      "put": {
        "tags": [
          "Jobs"
        ],
        "summary": "Hold Job",
        "responses": {
          "200": {
            "description": "Success"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "id"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "reason": "Awaiting documents"
              }
            }
          }
        }
      }
    },
    "/jobs/{id}/resume": {
      "put": {
        "tags": [
          "Jobs"
        ],
        "summary": "Resume Job",
        "responses": {
          "200": {
            "description": "Success"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "id"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "reason": "Documents received"
              }
            }
          }
        }
      }
    },
    "/jobs/{id}/clone": {
      "post": {
        "tags": [
          "Jobs"
        ],
        "summary": "Clone Job",
        "responses": {
          "200": {
            "description": "Success"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "id"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
...
    "/jobs/{id}": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "Get Job",
        "responses": {
          "200": {
            "description": "Success"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "id"
          }
        ]
      }
    },
    "/jobs/{id}/history": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "Job History",
        "responses": {
          "200": {
            "description": "Success"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "id"
          }
        ]
      }
    },
...
    "/client/jobs": {
      "get": {
        "tags": [
          "Client Portal"
        ],
        "summary": "Client Jobs",
        "responses": {
          "200": {
            "description": "Success"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/non-conformities/job/{jobId}": {
      "get": {
        "tags": [
          "Non-Conformities"
        ],
        "summary": "NC by Job",
        "responses": {
          "200": {
            "description": "Success"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "jobId"
          }
        ]
      }
    },
...
    "/mobile/offline/jobs": {
      "get": {
        "tags": [
          "Mobile"
        ],
        "summary": "Offline Jobs",
        "responses": {
          "200": {
            "description": "Success"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token from POST /auth/login"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}